+++ babs/__init__.py
@@ -4,3 +4,10 @@ try:
     from ._version import __version__
 except ImportError:
     __version__ = '0+unknown'
+
+from .bootstrap import BABSBootstrap
+from .check_setup import BABSCheckSetup
+from .interaction import BABSInteraction
+from .merge import BABSMerge
+
+__all__ = ['BABSBootstrap', 'BABSCheckSetup', 'BABSInteraction', 'BABSMerge']
