+++ docs/babs-init.rst
@@ -79,8 +79,8 @@ Examples are ``BIDS``, ``freesurfer``.
 How is the list of subjects (and sessions) determined?
 ------------------------------------------------------
 A list of subjects (and sessions) will be determined when running ``babs init``,
-and will be saved in a CSV file called named ``sub_final_inclu.csv`` (for single-session dataset)
-or ``sub_ses_final_inclu.csv`` (for multiple-session dataset),
+and will be saved in a CSV file called named ``processing_inclusion.csv`` (for single-session dataset)
+or ``processing_inclusion.csv`` (for multiple-session dataset),
 located at ``/path/to/my_BABS_project/analysis/code``.
 
 **To filter subjects and sessions**, use ``babs init`` with ``--list-sub-file /path/to/subject/list/csv/file``. Examples: `Single-session example <https://github.com/PennLINC/babs/blob/ba32e8fd2d6473466d3c33a1b17dfffc4438d541/notebooks/initial_sub_list_single-ses.csv>`_, `Multi-session example <https://github.com/PennLINC/babs/blob/ba32e8fd2d6473466d3c33a1b17dfffc4438d541/notebooks/initial_sub_list_multi-ses.csv>`_.
